добавляет leaflet на страницу выбранного оффера
