+++ src/components/list-offers/list-offers.tsx
@@ -1,19 +1,29 @@
+import cn from 'classnames';
 import { TOffer } from '../../types/offer';
-import CitiesCard from '../cities-card/card';
+import CitiesCard from '../offer-card/card';
 
 type TListOffers = {
   offers: TOffer[];
   onCardHover: (id: string | null) => void;
+  isMainOfferList?: boolean;
 };
 
-export default function ListOffers({ offers, onCardHover }: TListOffers): JSX.Element {
+export default function ListOffers({ offers, onCardHover, isMainOfferList }: TListOffers): JSX.Element {
   return (
-    <div className="cities__places-list places__list tabs__content">
+    <div
+      className={cn(
+        'places__list',
+        {'cities__places-list': isMainOfferList},
+        {'tabs__content': isMainOfferList},
+        {'near-places__list': !isMainOfferList}
+      )}
+    >
       {offers.map((offer) => (
         <CitiesCard
           key={offer.id}
           offer={offer}
           onCardHover={onCardHover}
+          isMainOfferList={isMainOfferList}
         />
       ))}
     </div>
