+++ src/components/offer-card/card.tsx
@@ -1,17 +1,24 @@
 import { Link } from 'react-router-dom';
+import cn from 'classnames';
 import { TOffer } from '../../types/offer';
 import { AppRoute, RATING_COEF } from '../../const';
-import BookmarksButton from '../../components/bookmark/favorite-bookmark';
+import BookmarksButton from '../bookmark/favorite-bookmark';
 
 type TCitiesCardProps = {
   offer: TOffer;
   onCardHover: (activeCard: string | null) => void;
+  isMainOfferList?: boolean;
 }
 
-function CitiesCard({offer, onCardHover}: TCitiesCardProps): JSX.Element {
+function CitiesCard({offer, onCardHover, isMainOfferList}: TCitiesCardProps): JSX.Element {
 
   return (
-    <article className="cities__card place-card"
+    <article
+      className={cn(
+        'place-card',
+        {'cities__card': isMainOfferList},
+        {'near-places__card': !isMainOfferList}
+      )}
       onMouseOver={() => onCardHover(offer.id)}
       onMouseOut={() => onCardHover(null)}
     >
@@ -20,16 +27,15 @@ function CitiesCard({offer, onCardHover}: TCitiesCardProps): JSX.Element {
           <span>Premium</span>
         </div>
       )}
-      <div className="cities__image-wrapper place-card__image-wrapper">
-        <a href="#">
-          <img
-            className="place-card__image"
-            src={offer.previewImage}
-            width="260"
-            height="200"
-            alt="Place image"
-          />
-        </a>
+      <div
+        className={cn(
+          'place-card__image-wrapper',
+          {'cities__image-wrapper': isMainOfferList},
+          {'near-places__image-wrapper': !isMainOfferList}
+        )}
+      >
+        <img className="place-card__image" src={offer.previewImage} width="260" height="200" alt={offer.title} />
+
       </div>
       <div className="place-card__info">
         <div className="place-card__price-wrapper">
