--- src/pages/offer/offer-page.tsx
@@ -1,291 +1,0 @@
-import {Helmet} from 'react-helmet-async';
-import { Navigate } from 'react-router-dom';
-import { useParams } from 'react-router-dom';
-import Header from '../../components/header/header';
-import OfferReview from '../../components/offers-review/review';
-import BookmarksButton from '../../components/bookmark/favorite-bookmark';
-import { TFullOffer } from '../../types/offer';
-import { TReview } from '../../types/review';
-import { AppRoute, RATING_COEF } from '../../const';
-
-type TOfferPageProps = {
-  fullOffers: TFullOffer[];
-  reviews: TReview[];
-};
-
-function OfferPage({fullOffers, reviews}: TOfferPageProps): JSX.Element {
-  const {id } = useParams();
-  const selectedOffer = fullOffers.find((offer) => offer.id === id);
-
-  if (!selectedOffer) {
-    return <Navigate to={AppRoute.NotFound}/>;
-  }
-
-  return (
-    <div className="page">
-      <Helmet>
-        <title>{`6 cities - ${selectedOffer.title}`}</title>
-      </Helmet>
-      <Header />
-      <main className="page__main page__main--offer">
-        <section className="offer">
-          <div className="offer__gallery-container container">
-            <div className="offer__gallery">
-              {selectedOffer &&
-                selectedOffer.images.map((item) => (
-                  <div
-                    className="offer__image-wrapper"
-                    key={crypto.randomUUID()}
-                  >
-                    <img
-                      className="offer__image"
-                      src={item}
-                      alt="Photo studio"
-                    />
-                  </div>
-                ))}
-            </div>
-          </div>
-          <div className="offer__container container">
-            <div className="offer__wrapper">
-              {selectedOffer && selectedOffer.isPremium && (
-                <div className="offer__mark">
-                  <span>Premium</span>
-                </div>
-              )}
-              <div className="offer__name-wrapper">
-                <h1 className="offer__name">{selectedOffer && selectedOffer.title}</h1>
-                <BookmarksButton
-                  isActive={false}
-                  size="big"
-                  page="offer"
-                />
-              </div>
-              <div className="offer__rating rating">
-                <div className="offer__stars rating__stars">
-                  <span style={{ width: `${Math.round(selectedOffer.rating) * RATING_COEF}%` }}></span>
-                  <span className="visually-hidden">Rating</span>
-                </div>
-                <span className="offer__rating-value rating__value">{selectedOffer.rating}</span>
-              </div>
-              <ul className="offer__features">
-                <li className="offer__feature offer__feature--entire">
-                  {selectedOffer && selectedOffer.type}
-                </li>
-                <li className="offer__feature offer__feature--bedrooms">
-                  {selectedOffer && selectedOffer.bedrooms} Bedrooms
-                </li>
-                <li className="offer__feature offer__feature--adults">
-                  Max {selectedOffer && selectedOffer.maxAdults} adults
-                </li>
-              </ul>
-              <div className="offer__price">
-                <b className="offer__price-value">&euro;{selectedOffer && selectedOffer.price}</b>
-                <span className="offer__price-text">&nbsp;night</span>
-              </div>
-              <div className="offer__inside">
-                <h2 className="offer__inside-title">What&apos;s inside</h2>
-                <ul className="offer__inside-list">
-                  {selectedOffer &&
-                    selectedOffer.goods.map((item) => (
-                      <li
-                        className="offer__inside-item"
-                        key={crypto.randomUUID()}
-                      >
-                        {item}
-                      </li>
-                    ))}
-                </ul>
-              </div>
-              <div className="offer__host">
-                <h2 className="offer__host-title">Meet the host</h2>
-                <div className="offer__host-user user">
-                  <div
-                    className={`offer__avatar-wrapper offer__avatar-wrapper${
-                      selectedOffer && selectedOffer.host.isPro ? '--pro' : ''
-                    } user__avatar-wrapper`}
-                  >
-                    <img
-                      className="offer__avatar user__avatar"
-                      src={selectedOffer && selectedOffer.host.avatarUrl}
-                      width="74"
-                      height="74"
-                      alt="Host avatar"
-                    />
-                  </div>
-                  <span className="offer__user-name">
-                    {selectedOffer && selectedOffer.host.name}
-                  </span>
-                  <span className="offer__user-status">
-                    {selectedOffer && selectedOffer.host.isPro ? 'Pro' : ''}
-                  </span>
-                </div>
-                <div className="offer__description">
-                  <p className="offer__text">
-                    {selectedOffer && selectedOffer.description}
-                  </p>
-                </div>
-              </div>
-              <OfferReview reviews={reviews}/>
-            </div>
-          </div>
-          <section className="offer__map map"></section>
-        </section>
-        <div className="container">
-          <section className="near-places places">
-            <h2 className="near-places__title">
-              Other places in the neighbourhood
-            </h2>
-            <div className="near-places__list places__list">
-              <article className="near-places__card place-card">
-                <div className="near-places__image-wrapper place-card__image-wrapper">
-                  <a href="#">
-                    <img
-                      className="place-card__image"
-                      src="img/room.jpg"
-                      width="260"
-                      height="200"
-                      alt="Place image"
-                    />
-                  </a>
-                </div>
-                <div className="place-card__info">
-                  <div className="place-card__price-wrapper">
-                    <div className="place-card__price">
-                      <b className="place-card__price-value">&euro;80</b>
-                      <span className="place-card__price-text">
-                        &#47;&nbsp;night
-                      </span>
-                    </div>
-                    <button
-                      className="place-card__bookmark-button place-card__bookmark-button--active button"
-                      type="button"
-                    >
-                      <svg
-                        className="place-card__bookmark-icon"
-                        width="18"
-                        height="19"
-                      >
-                        <use xlinkHref="#icon-bookmark"></use>
-                      </svg>
-                      <span className="visually-hidden">In bookmarks</span>
-                    </button>
-                  </div>
-                  <div className="place-card__rating rating">
-                    <div className="place-card__stars rating__stars">
-                      <span style={{ width: '80%' }}></span>
-                      <span className="visually-hidden">Rating</span>
-                    </div>
-                  </div>
-                  <h2 className="place-card__name">
-                    <a href="#">Wood and stone place</a>
-                  </h2>
-                  <p className="place-card__type">Private room</p>
-                </div>
-              </article>
-
-              <article className="near-places__card place-card">
-                <div className="near-places__image-wrapper place-card__image-wrapper">
-                  <a href="#">
-                    <img
-                      className="place-card__image"
-                      src="img/apartment-02.jpg"
-                      width="260"
-                      height="200"
-                      alt="Place image"
-                    />
-                  </a>
-                </div>
-                <div className="place-card__info">
-                  <div className="place-card__price-wrapper">
-                    <div className="place-card__price">
-                      <b className="place-card__price-value">&euro;132</b>
-                      <span className="place-card__price-text">
-                        &#47;&nbsp;night
-                      </span>
-                    </div>
-                    <button
-                      className="place-card__bookmark-button button"
-                      type="button"
-                    >
-                      <svg
-                        className="place-card__bookmark-icon"
-                        width="18"
-                        height="19"
-                      >
-                        <use xlinkHref="#icon-bookmark"></use>
-                      </svg>
-                      <span className="visually-hidden">To bookmarks</span>
-                    </button>
-                  </div>
-                  <div className="place-card__rating rating">
-                    <div className="place-card__stars rating__stars">
-                      <span style={{ width: '80%' }}></span>
-                      <span className="visually-hidden">Rating</span>
-                    </div>
-                  </div>
-                  <h2 className="place-card__name">
-                    <a href="#">Canal View Prinsengracht</a>
-                  </h2>
-                  <p className="place-card__type">Apartment</p>
-                </div>
-              </article>
-
-              <article className="near-places__card place-card">
-                <div className="place-card__mark">
-                  <span>Premium</span>
-                </div>
-                <div className="near-places__image-wrapper place-card__image-wrapper">
-                  <a href="#">
-                    <img
-                      className="place-card__image"
-                      src="img/apartment-03.jpg"
-                      width="260"
-                      height="200"
-                      alt="Place image"
-                    />
-                  </a>
-                </div>
-                <div className="place-card__info">
-                  <div className="place-card__price-wrapper">
-                    <div className="place-card__price">
-                      <b className="place-card__price-value">&euro;180</b>
-                      <span className="place-card__price-text">
-                        &#47;&nbsp;night
-                      </span>
-                    </div>
-                    <button
-                      className="place-card__bookmark-button button"
-                      type="button"
-                    >
-                      <svg
-                        className="place-card__bookmark-icon"
-                        width="18"
-                        height="19"
-                      >
-                        <use xlinkHref="#icon-bookmark"></use>
-                      </svg>
-                      <span className="visually-hidden">To bookmarks</span>
-                    </button>
-                  </div>
-                  <div className="place-card__rating rating">
-                    <div className="place-card__stars rating__stars">
-                      <span style={{ width: '100%' }}></span>
-                      <span className="visually-hidden">Rating</span>
-                    </div>
-                  </div>
-                  <h2 className="place-card__name">
-                    <a href="#">Nice, cozy, warm big bed apartment</a>
-                  </h2>
-                  <p className="place-card__type">Apartment</p>
-                </div>
-              </article>
-            </div>
-          </section>
-        </div>
-      </main>
-    </div>
-  );
-}
-
-export default OfferPage;
